+++ Integrations/VS-Code/src/commands/index.ts
@@ -9,6 +9,7 @@ import { selectInstance } from './project/selectInstance';
 import { generateClasses } from './project/generateClasses';
 import { pullScripts } from './project/pullScripts';
 import { pushScripts } from './project/pushScripts';
+import { pushScript } from './project/pushScript';
 import { syncScripts } from './project/syncScripts';
 import { openProjectFolder } from './project/openProjectFolder';
 import { compareWithRemote } from './project/compareWithRemote';
@@ -24,6 +25,7 @@ export function registerCommands(context: vscode.ExtensionContext) {
   const generateClassesCmd = vscode.commands.registerCommand(CommandIds.COMMAND_GENERATE_CLASSES, generateClasses);
   const pullScriptsCmd = vscode.commands.registerCommand(CommandIds.COMMAND_PULL_SCRIPTS, pullScripts);
   const pushScriptsCmd = vscode.commands.registerCommand(CommandIds.COMMAND_PUSH_SCRIPTS, pushScripts);
+  const pushScriptCmd = vscode.commands.registerCommand(CommandIds.COMMAND_PUSH_SCRIPT, pushScript);
   const syncScriptsCmd = vscode.commands.registerCommand(CommandIds.COMMAND_SYNC_SCRIPTS, syncScripts);
   const openProjectFolderCmd = vscode.commands.registerCommand(CommandIds.COMMAND_OPEN_PROJECT_FOLDER, openProjectFolder);
   const compareWithRemoteCmd = vscode.commands.registerCommand(CommandIds.COMMAND_COMPARE_WITH_REMOTE, compareWithRemote);
@@ -39,6 +41,7 @@ export function registerCommands(context: vscode.ExtensionContext) {
     generateClassesCmd,
     pullScriptsCmd,
     pushScriptsCmd,
+    pushScriptCmd,
     syncScriptsCmd,
     openProjectFolderCmd,
     compareWithRemoteCmd,